refresh local file list after downloading a remote file
